# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Akure Airport Staff Cooperative Society API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions (Admin+)
      parameters:
        - in: query
          name: memberId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
            enum: [savings, loan_repayment, store_repayment]
        - in: query
          name: month
          schema:
            type: string
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/my:
    get:
      operationId: listMyTransactions
      tags: [transactions]
      summary: List current member's transactions
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [savings, loan_repayment, store_repayment]
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # ── EXCEL UPLOAD ─────────────────────────────────────────────────────────────
...
  /store/items:
    get:
      operationId: listStoreItems
      tags: [store]
      summary: List available store items
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreItem"
    post:
      operationId: createStoreItem
      tags: [store]
      summary: Add a store item (Admin+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
  /store/items/{id}:
    get:
      operationId: getStoreItem
      tags: [store]
      summary: Get a store item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
        "404":
          description: Not found
    patch:
      operationId: updateStoreItem
      tags: [store]
      summary: Update a store item (Admin+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreItemBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
    delete:
      operationId: deleteStoreItem
      tags: [store]
      summary: Delete a store item (Admin+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /store/purchases:
    get:
      operationId: listStorePurchases
      tags: [store]
      summary: List store purchases (Admin+)
      parameters:
        - in: query
          name: memberId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StorePurchase"
    post:
      operationId: createStorePurchase
      tags: [store]
      summary: Record a store purchase (Member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorePurchaseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorePurchase"
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        type:
          type: string
          enum: [savings, loan_repayment, store_repayment]
        amount:
          type: number
        description:
          type: ["string", "null"]
        uploadRecordId:
          type: ["integer", "null"]
        month:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - type
        - amount
        - createdAt

    # ── EXCEL UPLOAD ─────────────────────────────────────────────────────────────
...
    StoreItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageObjectPath:
          type: ["string", "null"]
        quantityAvailable:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - quantityAvailable
        - isAvailable
        - createdAt
        - updatedAt
    CreateStoreItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageObjectPath:
          type: string
        quantityAvailable:
          type: integer
      required: [name, price]
    UpdateStoreItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageObjectPath:
          type: string
        quantityAvailable:
          type: integer
        isAvailable:
          type: boolean
    StorePurchase:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        storeItemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        outstandingBalance:
          type: number
        status:
          type: string
          enum: [outstanding, partial, settled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - storeItemId
        - itemName
        - quantity
        - unitPrice
        - totalPrice
        - outstandingBalance
        - status
        - createdAt
    CreateStorePurchaseBody:
      type: object
      properties:
        storeItemId:
          type: integer
        quantity:
          type: integer
        payFromSavings:
          type: boolean
      required: [storeItemId, quantity]
...